accept display-only fix changes
we previously had separate logic for displaying these fix indicators and
displaying fixes. we were only setting UnsafeFixes::Enabled, which translated to
an Applicability of Unsafe, which is what was checked for this fixability icon.
Then, we were unconditionally showing fixes, which included display-only fixes.
I think using the actual applicability for both of these makes sense but makes
these small changes to our display-only fix snapshots. since users still can't
enable display-only fixes in the CLI, this will not be a user-facing change